requirements.txt
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
exceptiongroup==1.2.2
Faker==27.0.0
filelock==3.15.4
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector-python==9.0.0
ordered-set==4.1.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
setuptools==74.1.0
six==1.16.0
SQLAlchemy==2.0.35
tldextract==5.1.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.0
zope.event==5.0
zope.interface==7.0.1
